Added tests, change evaluations to use double, any numeric type will be implicitly converted to double

+++ src/main/scala/com/scalatrees/Node.scala
@@ -5,9 +5,9 @@ package com.scalatrees
  * CNode: Constant node, holds a constant value, can return its value
  * FNode: Function node, holds a functions and nodes as parameters, can return its evaluated function
  */
-sealed trait Node { def evaluate: Int }
-case class CNode(value: Int) extends Node() { def evaluate: Int = value }
-case class FNode(func: Operation, val children: List[Node]) extends Node {  def evaluate: Int = func.eval(children map {_.evaluate} ) }
+sealed trait Node { def evaluate: Double }
+case class CNode(value: Double) extends Node() { def evaluate: Double = value }
+case class FNode(func: Operation, val children: List[Node]) extends Node {  def evaluate: Double = func.eval(children map {_.evaluate} ) }
 
 /** Node Operations
  * 
@@ -16,11 +16,11 @@ case class FNode(func: Operation, val children: List[Node]) extends Node {  def 
  * Subtract: Take the difference of two nodes
  * Multiple: Multiple two nodes together
  */
-sealed trait Operation  { def eval(node: List[Int]): Int }
-case object Add extends Operation { def eval(params: List[Int]) = params.reduce(_+_) }
-case object Mod extends Operation { def eval(params: List[Int]) = params.reduce(_%_) }
-case object Subtract extends Operation { def eval(params: List[Int]) = params.reduce(_-_) }
-case object Multiply extends Operation { def eval(params: List[Int]) = params.reduce(_*_) }
+sealed trait Operation  { def eval(node: List[Double]): Double }
+case object Add extends Operation { def eval(params: List[Double]) = params.reduce(_+_) }
+case object Mod extends Operation { def eval(params: List[Double]) = params.reduce(_%_) }
+case object Subtract extends Operation { def eval(params: List[Double]) = params.reduce(_-_) }
+case object Multiply extends Operation { def eval(params: List[Double]) = params.reduce(_*_) }
 //Power
 //Min
 //Max
